move enterprise/internal/license to internal/license
