+++ enterprise/cmd/frontend/internal/auth/confauth/middleware_test.go
@@ -11,8 +11,8 @@ import (
 
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/auth"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/external/session"
-	"github.com/sourcegraph/sourcegraph/enterprise/internal/license"
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/licensing"
+	"github.com/sourcegraph/sourcegraph/internal/license"
 )
 
 func TestMiddleware(t *testing.T) {
