+++ enterprise/cmd/frontend/internal/batches/resolvers/BUILD.bazel
@@ -143,7 +143,6 @@ go_test(
         "//enterprise/internal/batches/testing",
         "//enterprise/internal/batches/types",
         "//enterprise/internal/database",
-        "//enterprise/internal/license",
         "//internal/actor",
         "//internal/api",
         "//internal/conf",
@@ -158,6 +157,7 @@ go_test(
         "//internal/gitserver/gitdomain",
         "//internal/gqlutil",
         "//internal/httptestutil",
+        "//internal/license",
         "//internal/licensing",
         "//internal/observation",
         "//internal/rbac",
