+++ enterprise/cmd/frontend/internal/dotcom/productsubscription/BUILD.bazel
@@ -26,7 +26,6 @@ go_library(
     deps = [
         "//cmd/frontend/graphqlbackend",
         "//cmd/frontend/graphqlbackend/graphqlutil",
-        "//enterprise/internal/license",
         "//enterprise/internal/productsubscription",
         "//internal/accesstoken:licensing",
         "//internal/auth",
@@ -42,6 +41,7 @@ go_library(
         "//internal/featureflag",
         "//internal/gqlutil",
         "//internal/hashutil",
+        "//internal/license",
         "//internal/redispool",
         "//internal/slack",
         "//internal/types",
@@ -84,7 +84,6 @@ go_test(
     ],
     deps = [
         "//cmd/frontend/graphqlbackend",
-        "//enterprise/internal/license",
         "//enterprise/internal/productsubscription",
         "//internal/accesstoken:licensing",
         "//internal/actor",
@@ -96,6 +95,7 @@ go_test(
         "//internal/errcode",
         "//internal/featureflag",
         "//internal/hashutil",
+        "//internal/license",
         "//internal/redispool",
         "//internal/slack",
         "//internal/timeutil",
