+++ enterprise/cmd/frontend/internal/licensing/enforcement/BUILD.bazel
@@ -39,11 +39,11 @@ go_test(
     deps = [
         "//cmd/frontend/envvar",
         "//enterprise/internal/cloud",
-        "//enterprise/internal/license",
         "//internal/actor",
         "//internal/auth",
         "//internal/database",
         "//internal/extsvc",
+        "//internal/license",
         "//internal/types",
         "@com_github_derision_test_go_mockgen//testutil/require",
         "@com_github_inconshreveable_log15//:log15",
