+++ internal/licensing/BUILD.bazel
@@ -19,7 +19,6 @@ go_library(
     importpath = "github.com/sourcegraph/sourcegraph/internal/licensing",
     visibility = ["//enterprise:__subpackages__"],
     deps = [
-        "//enterprise/internal/license",
         "//internal/accesstoken:licensing",
         "//internal/conf",
         "//internal/conf/conftypes",
@@ -27,6 +26,7 @@ go_library(
         "//internal/errcode",
         "//internal/goroutine",
         "//internal/httpcli",
+        "//internal/license",
         "//internal/redispool",
         "//lib/errors",
         "@com_github_derision_test_glock//:glock",
@@ -53,7 +53,7 @@ go_test(
         "requires-network",
     ],
     deps = [
-        "//enterprise/internal/license",
+        "//internal/license",
         "//internal/redispool",
         "//lib/pointers",
         "@com_github_derision_test_glock//:glock",
